templates/langgraph/content/pyproject.toml
[project]
name = "langgraph-agent-template"
version = "0.1.0"
description = "A container-ready config-based LangGraph  & Agentic AI"
authors = [{ name = "Theofanis Petkos", email = "thepetk@gmail.com" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "openai>=1.0.0",
    "pyyaml>=6.0.0",
    "streamlit>=1.38.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
